pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "playwright>=0.1.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.0",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.0",
    "rollbar>=1.3.0",
    "ruff>=0.11.13",
    "whitenoise>=6.9.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]

[tool.ruff]
lint.select = ["E", "F", "W", "B", "I"]
line-length = 88
exclude = ["**/migrations/**", 'settings.py']
